% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worker.R
\name{initialize_worker}
\alias{initialize_worker}
\title{Creates new shiny.worker object}
\usage{
initialize_worker()
}
\value{
shiny.worker object
}
\description{
Creates new shiny.worker object
}
\examples{
worker <- initialize_worker()
}
